+++ packages/app/src/App.tsx
@@ -17,8 +17,6 @@ import {
   techdocsPlugin,
   TechDocsReaderPage,
 } from '@backstage/plugin-techdocs';
-import { TechDocsAddons } from '@backstage/plugin-techdocs-react';
-import { ReportIssue } from '@backstage/plugin-techdocs-module-addons-contrib';
 import {
   DefaultProviderSettings,
   UserSettingsPage,
@@ -118,11 +116,7 @@ const routes = (
     <Route
       path="/docs/:namespace/:kind/:name/*"
       element={<TechDocsReaderPage />}
-    >
-      <TechDocsAddons>
-        <ReportIssue />
-      </TechDocsAddons>
-    </Route>
+    />
     <Route path="/create" element={<ScaffolderPage />} />
     <Route
       path="/catalog-import"