+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - GS plugin: Add "Cluster access" pane that explains cluster access in the cluster list.
 
+### Removed
+
+- Disable Report Issue addon for TechDocs plugin.
+
 ## [0.23.1] - 2024-06-06
 
 ### Fixes